\name{UCBGrading-package}
\alias{UCBGrading-package}
\alias{UCBGrading}
\docType{package}
\title{
\packageTitle{UCBGrading}
}
\description{
\packageDescription{UCBGrading}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UCBGrading}
\packageIndices{UCBGrading}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{UCBGrading}

Maintainer: \packageMaintainer{UCBGrading}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
